+++ onyx/crates/ember-core/src/route.rs
@@ -1,8 +1,4 @@
-use ember_http::{
-    method::Method,
-    request::Request,
-    response::Response,
-};
+use ember_http::{method::Method, request::Request, response::Response};
 
 pub type Handler = Box<dyn Fn(Request) -> Response + Send + Sync>;
 
@@ -14,11 +10,7 @@ pub struct Route {
 }
 
 impl Route {
-    pub fn new(
-        method: Method,
-        path: impl Into<String>,
-        handler: Handler,
-    ) -> Self {
+    pub fn new(method: Method, path: impl Into<String>, handler: Handler) -> Self {
         Self {
             method,
             path: path.into(),
@@ -27,30 +19,20 @@ impl Route {
     }
 }
 
-
 #[cfg(test)]
 mod tests {
     use super::*;
-    use ember_http::{
-        response::Response,
-        status::StatusCode,
-    };
+    use ember_http::{response::Response, status::StatusCode};
 
     fn home(_: Request) -> Response {
-        Response::new(StatusCode::Ok)
-            .body("Home")
+        Response::new(StatusCode::Ok).body("Home")
     }
 
     #[test]
     fn creates_route() {
-        let route = Route::new(
-            Method::Get,
-            "/",
-            Box::new(home),
-        );
+        let route = Route::new(Method::Get, "/", Box::new(home));
 
         assert_eq!(route.method, Method::Get);
         assert_eq!(route.path, "/");
     }
 }
-